fix: improve dashboard tables (#242)
- fixes sidebar current link highlights
- fixes breadcrumbs
- adds country status and duration columns
- adds organisational unit type column to contributions
- institution status columns (can be multiple, maybe have dedicated
screen for relations, like contributions?)
- add institution country column
- add working group duration column

+++ apps/knowledge-base/app/(app)/[locale]/(dashboard)/dashboard/administrator/newsletters/_components/newsletters-page.tsx
@@ -97,7 +97,7 @@ export function NewslettersPage(props: Readonly<NewslettersPageProps>): ReactNod
 								<TableCell>{item.settings.subject_line}</TableCell>
 								<TableCell>
 									{item.send_time
-										? format.dateTime(new Date(item.send_time), { dateStyle: "long" })
+										? format.dateTime(new Date(item.send_time), { dateStyle: "short" })
 										: null}
 								</TableCell>
 								<TableCell>{format.number(item.emails_sent)}</TableCell>
